refactor(metrics): split into package (config, instruments, providers, collection) with re-exports
Replace monolithic metrics module with modular package. Preserve API and semantics; unit tests pass.
