+++ src/meridian/observability/metrics/__init__.py
@@ -1,0 +1,20 @@
+from __future__ import annotations
+
+from .config import PrometheusConfig, DEFAULT_LATENCY_BUCKETS
+from .instruments import Counter, Gauge, Histogram, Metrics
+from .providers import PrometheusMetrics, NoopMetrics, get_metrics, configure_metrics
+from .collection import time_block
+
+__all__ = [
+    "PrometheusConfig",
+    "DEFAULT_LATENCY_BUCKETS",
+    "Counter",
+    "Gauge",
+    "Histogram",
+    "Metrics",
+    "PrometheusMetrics",
+    "NoopMetrics",
+    "get_metrics",
+    "configure_metrics",
+    "time_block",
+]
